Move the intercepting route to the lower level
