+++ app/next-client-app/app/scanreports/[id]/mapping_rules/@summary_modal/default.tsx
